Introducing pages e componentization to the code
This commit introduces a new `CreateTripPage` and `TripDetailsPage`, along with their necessary components, to build out the trip creation feature. The `CreateTripPage` page allows users to set a destination, date, and invite guests, while the `TripDetailsPage` is a placeholder for future trip details. Furthermore, the application router was updated to route these new pages.